changed the logic and, converted generate orders to generate cart also, changed the models

+++ core/views.py
@@ -1,3 +1,4 @@
+from django.db.models import Q
 from django.shortcuts import get_object_or_404
 from rest_framework import status
 from rest_framework.response import Response
@@ -7,17 +8,23 @@ from users.serializers import UserSerializer
 from core import models, serializers
 
 
-class OrderAPIView(APIView):
+class CartAPIView(APIView):
+    '''
+    GENERATE USER'S CART
+
+    ALLOWED METHODS: GET, PUT, DELETE
+    '''
+
     def get(self, request):
-        orders = get_object_or_404(models.Order, placed_by=request.user)
-        serializer = serializers.OrderSerializer(orders, many=False)
+        orders = get_object_or_404(models.Cart, placed_by=request.user)
+        serializer = serializers.CartSerializer(orders, many=False)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def put(self, request):
         data = {**request.data}
         data.update({'placed_by': {'email': request.user.email}})
 
-        serializer = serializers.OrderSerializer(data=data)
+        serializer = serializers.CartSerializer(data=data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -25,6 +32,20 @@ class OrderAPIView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request):
-        order = get_object_or_404(models.Order, placed_by=request.user)
+        order = get_object_or_404(models.Cart, placed_by=request.user)
+
+        if order.status == 'DELIVERED':
+            return Response(status=status.HTTP_204_NO_CONTENT)
+
+        if request.query_params.get('delete') == 'item':
+            if not request.query_params.get('id'):
+                return Response({'message': 'Item Id is missing'}, status=status.HTTP_400_BAD_REQUEST)
+            item = get_object_or_404(
+                models.CartDetails,
+                Q(details__placed_by=request.user) &
+                Q(id=request.query_params.get('id'))
+            )
+            item.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
         order.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
